Functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
basketball_reference_web_scraper==4.14.0
numpy==1.26.3
pandas==2.2.2